Use animator only when navigating multiple levels at once
When pushing pages using animatorPush the animation will be finished
before the page is ready. This causes some weird lag after the new
page has been activated and should already be usable. This might be
a bug in PageStack.

It is much faster than regular pushing, though, so we use it when
pushing multiple pages at once. This seems like a good tradeoff between
perceived responsiveness and performance.