Only translate "Norwegian", as "Bokmål" is a proper name
"Bokmål" is now on the value side of the contributors list. If we
get translations for (an)other Norwegian language(s), we'll group the
contributors accordingly. This helps with translating, as the name
"Bokmål" can't/shouldn't be translated into other languages (whereas
"Norwegian" can and should be). Problem: the form "foo: bar" is certainly
incorrect in some languages...

It's not perfect; it might be changed again.