Implement synchronised bookmarks
Bookmarks are now saved in a separate file $CONFIG/bookmarks.json,
and no longer in the main config file. Bookmarks are automatically
migrated to the new location if the new file does not exist.

Bookmarks are automatically reloaded when the file changes on disk.
When bookmarks are changed in the app, they are saved immediately.

This way, changes to the bookmarks are always synchronised between
multiple app windows.