en: full stop → dot
   In computing, it is called a dot.

Wikipedia is not *best* source of all information, however, in this case
under the `Computing`_ section of :t:`Full stop` article, the word “dot”
is prefered (such as ‘dot com’, ‘dotfiles’, & so on). As mentioned at the top
of the article, the word “period” is much more common in North American
English (which I am a speaker of) so the term “full stop” was unexpected
to see in the UI. Historically, ‘full stop’ was used for the terminating
punctuation; it is unfortunate that Unicode has chosen the name ``FULL
STOP``. Seeing as “Behavior” (:ac:`U.S.` spelling) is one section
heading it is rather inconsistent to be mixing dialects.

.. _Computing: https://en.wikipedia.org/wiki/Full_stop#Computing

Format: reStructuredText